lots of cool stuff

+++ src/Main.cpp
@@ -5,7 +5,7 @@
 
 int main(int argc, char** argv) {
     std::vector<std::string> args;
-    for(size_t i = 0; i < argc; i++)
+    for(int i = 0; i < argc; i++)
         args.push_back(argv[i]);
 
     Ui ui(args);
